move code to new script
