+++ Fragments Of You/Assets/Scripts/PinkMovement.cs
@@ -7,7 +7,6 @@ public class PinkMovement : MonoBehaviour
     private Rigidbody2D rb;
     private BoxCollider2D coll;
     private SpriteRenderer sprite;
-    private SpringJoint2D arm;
     private Animator animator;
 
     public GameOver_Condition gameOver_Condition;
@@ -15,13 +14,10 @@ public class PinkMovement : MonoBehaviour
     [SerializeField] private LayerMask jumpableGround;
     private Vector2 respawnPoint;
     private float dirX = 0f;
-    private bool isAnchored = false;
     [SerializeField] private float moveSpeed = 7f;
     [SerializeField] private float jumpForce = 14f;
-    [SerializeField] private float grappleJumpForce = 7f;
-    [SerializeField] private float anchorableDis = 14f;
 
-    private bool hasArms = true;
+    //private bool hasArms = false;
     private bool hasLegs = true;
 
     private void Start()
@@ -30,14 +26,10 @@ public class PinkMovement : MonoBehaviour
         rb = GetComponent<Rigidbody2D>();
         coll = GetComponent<BoxCollider2D>();
         sprite = GetComponent<SpriteRenderer>();
-        arm = GetComponent<SpringJoint2D>();
         animator = GetComponent<Animator>();
 
         // freeze rotation
         rb.freezeRotation = true;
-
-        arm.enabled = false;
-        isAnchored = false;
 
         // set respawn to starting position
         respawnPoint = this.transform.position;
@@ -58,20 +50,7 @@ public class PinkMovement : MonoBehaviour
             {
                 Jump();
             }
-            else if(IsAnyValidAnchor() && !isAnchored && hasArms)
-            {
-                Grapple();
-            } else if(isAnchored)
-            {
-                GrappleOff();
-            }
         }
-
-        if(IsAnyValidAnchor())
-        {
-            arm.connectedBody = FindValidAnchor().GetComponent<Rigidbody2D>();
-        }
-        AnchorRadar();
 
         // if player falls off level respawn
         if(this.transform.position.y < -8)
@@ -103,78 +82,12 @@ public class PinkMovement : MonoBehaviour
         animator.SetBool("isJumping", true);
     }
 
+    // GroundCheck
     private bool IsGrounded()
     {
         return Physics2D.BoxCast(coll.bounds.center, coll.bounds.size, 0f, Vector2.down, .1f, jumpableGround);
     }
 
-    // Grapple
-    private void Grapple()
-    {
-        arm.enabled = true;
-        isAnchored = true;
-        animator.SetBool("isJumping", true);
-    }
-    // Disengage Grapple
-    private void GrappleOff()
-    {
-        arm.enabled = false;
-        isAnchored = false;
-        rb.AddForce(Vector2.up * grappleJumpForce, ForceMode2D.Impulse);
-    }
-
-    // Find Valid Anchor within anchorableDis
-    private GameObject FindValidAnchor() 
-    {
-        GameObject[] anchors;
-        anchors = GameObject.FindGameObjectsWithTag("Anchor");
-        GameObject valid = null;
-        float distance = anchorableDis;
-        Vector3 position = transform.position;
-        foreach (GameObject an in anchors)
-        {
-            Vector3 diff = an.transform.position - position;
-            float curDistance = diff.sqrMagnitude;
-            if (curDistance < distance)
-            {
-                valid = an;
-                distance = curDistance;
-            }
-        }
-        return valid;
-    }
-    // Return true if an anchor within range
-    private bool IsAnyValidAnchor() 
-    {
-        GameObject[] anchors;
-        anchors = GameObject.FindGameObjectsWithTag("Anchor");
-        float distance = anchorableDis;
-        Vector3 position = transform.position;
-        foreach (GameObject an in anchors)
-        {
-            Vector3 diff = an.transform.position - position;
-            float curDistance = diff.sqrMagnitude;
-            if (curDistance < distance)
-            {
-                return true;
-            }
-        }
-        return false;
-    }
-    
-    private void AnchorRadar()
-    {
-        GameObject an = FindValidAnchor();
-        if(IsAnyValidAnchor()) {
-            if(isAnchored) {
-                Debug.DrawLine(transform.position, an.transform.position, Color.green);
-            } else 
-            {
-                Debug.DrawLine(transform.position, an.transform.position, Color.red);
-            }
-            
-        }
-    }
     // Grapple functions end ----------------------------------------------------------------------
 
     // Respawn functions start --------------------------------------------------------------------
@@ -203,7 +116,7 @@ public class PinkMovement : MonoBehaviour
     // function to lose arms
     public void loseArms()
     {
-        hasArms = false;
+        //hasArms = false;
         // preform sprite switch here
 
     }
